Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 140ms, Memory: 8MB, Status: Passed - CodeTree

+++ 241009/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -2,10 +2,10 @@ import java.util.*;
 import java.io.*;
 
 public class Main {
-    static int m;
-    static int a,b;
-    static int minResult = Integer.MAX_VALUE;
-    static int maxResult = Integer.MIN_VALUE;
+    static long m;
+    static long a,b;
+    static long minResult = Long.MAX_VALUE;
+    static long maxResult = Long.MIN_VALUE;
     
     public static void main(String[] args) throws IOException{
         // 여기에 코드를 작성해주세요.
@@ -15,16 +15,16 @@ public class Main {
 
     private static void init() throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        m = Integer.parseInt(br.readLine());
+        m = Long.parseLong(br.readLine());
 
         StringTokenizer st = new StringTokenizer(br.readLine());
-        a = Integer.parseInt(st.nextToken());
-        b = Integer.parseInt(st.nextToken());
+        a = Long.parseLong(st.nextToken());
+        b = Long.parseLong(st.nextToken());
     }
 
     private static void simulate() {
-        for(int i = a; i<=b; i++) {
-            int count = find(i);
+        for(long i = a; i<=b; i++) {
+            long count = find(i);
             minResult = Math.min(minResult,count);
             maxResult = Math.max(maxResult,count);
         }
@@ -32,13 +32,13 @@ public class Main {
         System.out.println(minResult +" "+maxResult);
     }
 
-    private static int find(int target) {
-        int left = 1;
-        int right = m;
+    private static int find(long target) {
+        long left = 1;
+        long right = m;
         int count = 0;
         while(left<=right) {
             count++;
-            int mid = (left+right)/2;
+            long mid = (left+right)/2;
             if(mid > target) {
                 right = mid -1;
             }else if (mid<target) {
